Fix Nearby API
- Move Nearby instance method signatures to ExposureNotificationClient
- Make ExposureNotificationClient an interface
- Return an ExposureNotificationClient instance in Nearby.getENC()
